+++ roseau/load_flow/models/loads.py
@@ -209,6 +209,7 @@ class PowerLoad(AbstractLoad["CyPowerLoad | CyDeltaPowerLoad | CyFlexibleLoad | 
     """A constant power load."""
 
     type: Final = "power"
+    _short_circuit_compatible: Final = False
 
     def __init__(
         self,
@@ -263,14 +264,6 @@ class PowerLoad(AbstractLoad["CyPowerLoad | CyDeltaPowerLoad | CyFlexibleLoad | 
                 by default. To override the default behavior, pass an explicit ``True`` or ``False``.
         """
         super().__init__(id=id, bus=bus, phases=phases, connect_neutral=connect_neutral)
-
-        if bus.short_circuits:
-            msg = (
-                f"The power load {self.id!r} is connected on bus {bus.id!r} that already has a short-circuit. "
-                f"It makes the short-circuit calculation impossible."
-            )
-            logger.error(msg)
-            raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_SHORT_CIRCUIT)
         if flexible_params is not None:
             if isinstance(flexible_params, FlexibleParameter):
                 flexible_params = [flexible_params] * self._size
@@ -386,6 +379,7 @@ class CurrentLoad(AbstractLoad["CyCurrentLoad | CyDeltaCurrentLoad"]):
     """A constant current load."""
 
     type: Final = "current"
+    _short_circuit_compatible: Final = False
 
     def __init__(self, id: Id, bus: Bus, *, currents: ComplexScalarOrArrayLike1D, phases: str | None = None) -> None:
         """CurrentLoad constructor.
@@ -415,14 +409,6 @@ class CurrentLoad(AbstractLoad["CyCurrentLoad | CyDeltaCurrentLoad"]):
                 be present in the phases of the connected bus.
         """
         super().__init__(id=id, phases=phases, bus=bus)
-
-        if bus.short_circuits:
-            msg = (
-                f"The current load {self.id!r} is connected on bus {bus.id!r} that already has a short-circuit. "
-                f"It makes the short-circuit calculation impossible."
-            )
-            logger.error(msg)
-            raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_SHORT_CIRCUIT)
         if self.has_floating_neutral:
             msg = (
                 f"Constant current loads cannot have a floating neutral. {type(self).__name__} "
@@ -461,6 +447,7 @@ class ImpedanceLoad(AbstractLoad["CyAdmittanceLoad | CyDeltaAdmittanceLoad"]):
     """A constant impedance load."""
 
     type: Final = "impedance"
+    _short_circuit_compatible: Final = True
 
     def __init__(
         self,